makerworks-backend/requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
aiosqlite==0.21.0
bcrypt==4.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3

# ✅ Ensure gunicorn is installed for runtime
gunicorn==23.0.0

h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
nvidia-ml-py3==7.352.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
stripe==12.3.0
typing-inspection==0.4.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
itsdangerous>=2.1.2
numpy==2.3.1
trimesh==4.7.1
imageio[ffmpeg]==2.35.1
pyglet<2

# ✅ Keep plain PyOpenGL for rendering, remove accelerate to avoid 3.12+ build errors
PyOpenGL==3.1.0
pyrender==0.1.45
